Version 0.27 - October 2013
1. Modifying the mergeNodes routine to make it cleaner and easier to handle cases
of option nodes associated with filename stubs.  When there are filename stubs, there
can be multiple output files associated with the option node with different names. If
another task uses these files, there has to be a decision about which of the file nodes
are merged and which are left alone.