Version 0.150 - October 2014
1. Added the 'additional nodes' section to pipeline configuration files and processing
steps to handle them. This allows additional nodes to be added to tasks that may well
have the same argument as a previously defined node. This is required for the following
situation:

1. TaskA has 'FileA' as input, set with argument ArgumentA.
2. TaskB has 'FileA' and 'FileB' as input, both set with ArgumentB.

A graph node is created and a pipeline argument assigned and this node is
accessed by both TaskA and TaskB to assign FileA. Assigning FileB to this
pipeline argument will force both TaskA and TaskB to accept both files.

Now a new node is added that is only connected to TaskB. TaskB now has two
nodes using the same argument, but the new node is isolated from TaskA,
allowing both TaskA and TaskB to get FileA from the same node and FileB is
accessed by Task B only.